This commit designates two changes: MIT and deprecation of Qt5 interface.
As I'm the only code contributor, I'm announcing that URTrator code
is now licensed under MIT license.

Also I'm deprecating Qt5 interface as it is really hard to write Qt
code and create installers with it. Maybe one day I even figure out
how to work with QtQuick (which attracts me more and more ATM) and
URTrator will use it, but not now.